First version of performance measurement rig
Runs iperf3 via wireguard-go, main branch of NepTUN and the current
branch between two docker containers. The results are printed in the
job log. Probably we want to add some configurability to those scripts
but it is better to do it as we go, when we actually know what we need.